+++ server/UserData/User/unlocks.cs
@@ -2,7 +2,7 @@ namespace Nanina.UserData
 {
     public class Unlocks
     {
-        public byte maxDungeonFloor { get; set; } = 1;
+        public short maxDungeonFloor { get; set; } = 1;
 
     }
 }
